assigning workouts to users and app was successfully completed

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,13 +1,20 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) return;
+
     const res = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        "Authorization": `Bearer ${user.token}`,
+      },
     });
 
     const data = await res.json();
